backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==3.2
django-cors-headers==4.3.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
idna==3.4
oauthlib==3.2.2
Pillow==10.1.0
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.6

catboost==1.2.2
gensim==4.3.2
geopy==2.4.0
joblib==1.3.2
keras==2.14.0
Keras-Preprocessing==1.1.2
numpy==1.24.1
pandas==2.0.3
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
scikit-learn==1.3.1
scipy==1.10.1
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.34.0
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
nltk==3.8.1